requirements.txt
google-genai>=0.3.0
aiogram>=3.0.0
redis>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy
yt-dlp
trafilatura
playwright
playwright-stealth
apscheduler
pytz
